test(6kyu): multiple 3 or 5
